add new storybook for Notification component (#282)
